src/app/web-authn.service.spec.ts
import { TestBed } from '@angular/core/testing';

import { WebAuthnService } from './web-authn.service';
import { HttpClient } from '@angular/common/http';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';

describe('WebAuthnService', () => {
  let httpClient: HttpClient;
  let httpTestingController: HttpTestingController;
  let service: WebAuthnService;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule],
      providers: [WebAuthnService]
    });

    httpClient = TestBed.inject(HttpClient);
    httpTestingController = TestBed.inject(HttpTestingController);
    service = TestBed.inject(WebAuthnService);
  });

  afterEach(() => {
    // After every test, assert that there are no more pending requests.
    httpTestingController.verify();
  });

  it('should be created', () => {
    expect(service).toBeTruthy();
  });
});
